Add ability to rebuild binaries on pushes
When a webhook message arrives with an update to the `master_branch`,
`fisherman` will now build a release version of the updated binary by
executing `cargo build --release` in the directory for the project.

This is specific to Cargo projects and relies on the directory
structure, so is not very customisable. An attempt was made to use the
`cargo` library itself instead of spawning a `Command`, but it failed to
build on the server due to memory issues.