Allow configuration options to be specified
Currently, most paths are hardcoded based on the server that runs the
project, meaning it can be somewhat difficult to develop it locally.
This allows options such as the repository directory, SSH private key
file and Cargo binary location to be specified, meaning they can be
changed easily based on the system running `fisherman`.