Allow different code roots and multiple binaries
Add support for changing the code root of a specific repository,
allowing you to specify that your Rust code lives in `/backend` instead
of `/` as is default.

Additionally, allow users to provide a list of all the binaries to build
on changes to the project. Each of these will then be built and
restarted within Supervisor.