Allow webhooks to be verified through secrets
GitHub provides the option for a shared secret when creating a webhook,
allowing them to be signed through HMAC-SHA256 and the hash included in
the headers. This allows us to verify that the payload is from GitHub,
as otherwise malicious users could simply send valid JSON to the webhook
and attempt to exploit it.

This firstly requires us to allow secrets as an option in the
configuration. A repository's specific secret will be used if that
exists, followed by the global secret value. If no secret is set, the
webhook will only be allowed for processing if it also does not contain
a HMAC hash.

Additionally, JSON is first read as a stream of bytes and then converted
into JSON using `serde_json`. This is so that we can ensure the raw
bytes are as expected.

Closes #1.