Allow branches other than `master` to be followed
Previously, `fisherman` would ignore all events other than that of the
`master` branch, meaning users could not use the project to track
changes to a development server for example.

This allows them to specify a `follow` key for each repository, defining
the changes that `fisherman` should care about. If no `follow` key is
specified, it will revert to the default behaviour and follow the master
branch.

Closes #16.