Allow `fisherman` to respond to ping messages
When a webhook is setup through the GitHub website, a ping message will
be sent to the provided URL. Currently, `fisherman` will panic here, as
it will fail to deserialize it as a valid push message and thus respond
with a 500/502.

Instead, this allows `fisherman` to deserialize as either a ping or push
message depending on the value of the `X-GitHub-Event` header and then
handle either path. Push events are handled just as before, whereas a
ping event will immediately respond with the name of the repository and
the URL where `fisherman` is listening.

This means users can setup `fisherman` on their server and then setup
the webhook in the browser to get instant feedback in whether their
instance is working correctly. If not, they can change the settings and
resend the payload until it responds with the correct message and status
code.

Closes #20.