feat: basic status page
Add a basic status page that shows some of the events that have occurred
in the system. This is done through a shared state between threads,
which contains a timeseries queue of event data. When querying
`/status`, a list of timestamps and their related events will be
displayed.

Currently, this only occurs on `ping` events, as refactoring of the
queue needs to occur before it can be done for `push` events which are
more important.