feat: alert on failed builds
When `fisherman` fails to update a production instance of a repository,
send a message to the relevant Discord channel explaining the reason for
this if possible. This allows the developer to be aware that something
went wrong during deployment and that they should go and fix it.

Error messages will be sent if `fisherman` fails to pull changes, build
a binary, restart the process within `supervisor` or run any of the
additional commands specified in the configuration.

This operation is short-circuiting and has no rollback, meaning
`fisherman` will not continue to build or restart other binaries if a
previous operation failed. It will also not attempt to rollback to a
previously working version of the repository.

Closes #31.