feat: allow `fisherman` to run precommands
`fisherman` already supports running commands after processes have been
restarted, which can be used to build the frontend of a JavaScript
application for example. This allows precommands to be specified, which
are executed immediately after the changes are pulled and before any
binaries are rebuilt. This is aimed at processes such as database
migrations.